requirements.txt
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.0
google-api-python-client==2.102.0
pandas==2.0.3
lxml==4.9.3